feat(app): removing utilities.ts file
[ci skip]
